requirements.txt
Pillow==9.0.0
python-dotenv==0.19.2